+++ packages/fiber/src/index.tsx
@@ -2,8 +2,9 @@
 export * from './three-types'
 import * as ReactThreeFiber from './three-types'
 export { ReactThreeFiber }
+export type { BaseInstance, LocalState } from './core/renderer'
 export type {
-  Intersection,
+  IIntersection as Intersection,
   Subscription,
   Dpr,
   Size,